requirements.txt
Flask
flask-session
requests
gTTS
geopy
fuzzywuzzy
SpeechRecognition
python-dotenv